personal_assistant/requirements.txt

streamlit
O365
python-dotenv
langchain-openai
langchain-google-vertexai
beautifulsoup4
dateparser
